Pixel split experiment working nicely

+++ src/main/java/com/tgelder/downhill/geometry/Scale.java
@@ -2,27 +2,27 @@ package com.tgelder.downhill.geometry;
 
 public class Scale {
   
-  private float outMin;
-  private float inMin;
+  private double outMin;
+  private double inMin;
 
-  private float factor;
+  private double factor;
   
   public Scale() {
     
   }
   
-  public Scale(float inMin, float inMax, float outMin, float outMax) {
+  public Scale(double inMin, double inMax, double outMin, double outMax) {
     set(inMin, inMax, outMin, outMax);
   }
   
-  public void set(float inMin, float inMax, float outMin, float outMax) {
+  public void set(double inMin, double inMax, double outMin, double outMax) {
     this.inMin = inMin;
     this.outMin = outMin;
     
     factor = (outMax - outMin)/(inMax - inMin);
   }
   
-  public float scale(float value) {
+  public double scale(double value) {
     return (value - inMin)*factor + outMin;
   }
